refactor: move `capitalize` util function to `string.ts`
