+++ apps/calendar/src/utils/keyboard.ts
@@ -4,7 +4,3 @@ import { KEYCODE } from '@src/constants/keyboard';
 export function isKeyPressed(e: KeyboardEvent, key: KEY) {
   return e.key ? e.key === key : e.keyCode === KEYCODE[key];
 }
-
-export function capitalize(str: string) {
-  return str.charAt(0).toUpperCase() + str.slice(1);
-}
