+++ apps/calendar/src/utils/string.ts
@@ -1,0 +1,3 @@
+export function capitalize(str: string) {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
